Fix CI warnings
* libwget/css.c (wget_css_parse_buffer): Explicitly typecaste 'len' to int from size_t
* libwget/cookie.c (_compare_cookie2): Change datatype of 'n' from int to size_t
* libwget/cookie.c (_compare_cookie2): Typecaste 'n' to int explicitly
* libwget/http.c (_send_callback): Change datatype of 'rc' from int to ssize_t
* libwget/http.c (_recv_callback): Likewise
* libwget/http.c (wget_http_match_no_proxy): Change param name from 'no_proxies' to
  'no_proxies_vec' to not to shadow global 'no_proxies'
* libwget/ssl_gnutls.c (check_ocsp_response): Explicit type conversion to unsigned int
* libwget/ssl_gnutls.c (check_ocsp_response): Delete non-reachable code
* libwget/net.c (wget_tcp_listen): Remove goto
* libwget/net.c (wget_tcp_listen): Move declaration of variable 'on' in #ifdef block where it's used
* src/options.c (set_long_option): Change datatypes of 'invert', 'char_insensitive'
  and 'value_present' from int to char

+++ libwget/cookie.c
@@ -136,7 +136,7 @@ static int G_GNUC_WGET_NONNULL_ALL G_GNUC_WGET_PURE _compare_cookie(const wget_c
 // this is how we sort the entries when constructing a Cookie: header field
 static int G_GNUC_WGET_NONNULL_ALL G_GNUC_WGET_PURE _compare_cookie2(const wget_cookie_t *c1, const wget_cookie_t *c2)
 {
-	int n;
+	size_t n;
 
 	// RFC 6265 5.4 demands sorting by 1. longer paths first, 2. earlier creation time first.
 
@@ -145,7 +145,7 @@ static int G_GNUC_WGET_NONNULL_ALL G_GNUC_WGET_PURE _compare_cookie2(const wget_
 		// n = c1->creation - c2->creation;
 	}
 
-	return n;
+	return (int)n;
 }
 
 static int G_GNUC_WGET_NONNULL_ALL _domain_match(const char *domain, const char *host)
